Add unit test for Charges

+++ src/Game/NeoServer.Game.Common/Contracts/Items/IChargeable.cs
@@ -2,7 +2,7 @@
 {
     public interface IChargeable
     {
-        byte Charges { get; }
+        ushort Charges { get; }
         void DecreaseCharges();
     }
 }
